Long de Proveedores

+++ LubriTech/Model/Supplier_Information/Supplier.cs
@@ -12,9 +12,9 @@ namespace LubriTech.Model.Supplier_Information
         public string id { get; set; }
         public string name { get; set; }
         public string email { get; set; }
-        public int phone { get; set; }
+        public long phone { get; set; }
 
-        public Supplier(string id, string name, string email, int phone)
+        public Supplier(string id, string name, string email, long phone)
         {
             this.id = id;
             this.name = name;
