+++ LubriTech/Model/Supplier_Information/Supplier_Model.cs
@@ -35,7 +35,7 @@ namespace LubriTech.Model.Supplier_Information
 
                 while (reader.Read())
                     {
-                        suppliers.Add(new Supplier(reader["Identificacion"].ToString(), reader["Nombre"].ToString(), reader["CorreoElectronico"].ToString(), Convert.ToInt32(reader["NumeroTelefono"])));
+                        suppliers.Add(new Supplier(reader["Identificacion"].ToString(), reader["Nombre"].ToString(), reader["CorreoElectronico"].ToString(), Convert.ToInt64(reader["NumeroTelefono"])));
                     }
                     return suppliers;
                 }
@@ -69,7 +69,7 @@ namespace LubriTech.Model.Supplier_Information
 
                 while (reader.Read())
                 {
-                    supplier = new Supplier(reader["Identificacion"].ToString(), reader["Nombre"].ToString(), reader["CorreoElectronico"].ToString(), Convert.ToInt32(reader["NumeroTelefono"]));
+                    supplier = new Supplier(reader["Identificacion"].ToString(), reader["Nombre"].ToString(), reader["CorreoElectronico"].ToString(), Convert.ToInt64(reader["NumeroTelefono"]));
                 }
                 return supplier;
             }
